Add files via upload
An artifact has an incremental number (index), a creator (creator), and a year of creation (year) that are assigned to it from the beginning and remain constant. Additionally, it has a function (getInfo) that prints the creator and the year of creation of the artifact, as well as a function (getIndex) that prints only the incremental number.